*** empty log message ***

+++ magick/transform.c
@@ -1334,8 +1334,8 @@ MagickExport Image *RollImage(const Image *image,const long x_offset,
     indexes=GetIndexes(image);
     for (x=0; x < (long) image->columns; x++)
     {
-      q=SetImagePixels(roll_image,(offset.x+x) % image->columns,
-        (offset.y+y) % image->rows,1,1);
+      q=SetImagePixels(roll_image,(long) (offset.x+x) % image->columns,
+        (long) (offset.y+y) % image->rows,1,1);
       if (q == (PixelPacket *) NULL)
         break;
       roll_indexes=GetIndexes(roll_image);
